Extract copy address input to separate component
