+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Contests/CopyAddressInput/CopyAddressInput.scss
@@ -1,0 +1,18 @@
+@import '../../../../../../styles/shared';
+
+.CopyAddressInput {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  padding: 5px 12px;
+  background-color: $neutral-50;
+  border: 1px solid $neutral-200;
+  border-radius: 6px;
+
+  .address {
+    margin-right: 12px;
+    word-break: break-all;
+    text-align: left;
+  }
+}
