+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Contests/CopyAddressInput/index.ts
@@ -1,0 +1,3 @@
+import CopyAddressInput from './CopyAddressInput';
+
+export default CopyAddressInput;
