+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Contests/ManageContest/steps/ContestLiveStep/ContestLiveStep.scss
@@ -27,23 +27,6 @@
     justify-content: center;
     gap: 26px;
 
-    .copy-address-input {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      width: 100%;
-      padding: 5px 12px;
-      background-color: $neutral-50;
-      border: 1px solid $neutral-200;
-      border-radius: 6px;
-
-      .address {
-        margin-right: 12px;
-        word-break: break-all;
-        text-align: left;
-      }
-    }
-
     .description {
       color: $neutral-500;
     }
