+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Contests/ManageContest/steps/ContestLiveStep/ContestLiveStep.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 
 import { useAnimation } from 'hooks/useAnimation';
 import { useCommonNavigate } from 'navigation/helpers';
-import { saveToClipboard } from 'utils/clipboard';
 import { CWText } from 'views/components/component_kit/cw_text';
 import { CWButton } from 'views/components/component_kit/new_designs/CWButton';
-import CWIconButton from 'views/components/component_kit/new_designs/CWIconButton';
 import CWPageLayout from 'views/components/component_kit/new_designs/CWPageLayout';
+
+import CopyAddressInput from '../../../CopyAddressInput';
 
 import './ContestLiveStep.scss';
 
@@ -15,10 +15,6 @@ const ContestLiveStep = () => {
   const { animationStyles } = useAnimation();
 
   const createdContestAddress = '0xb794f5ea0ba39494ce839613fffba74279579268';
-
-  const handleCopyLink = async () => {
-    await saveToClipboard(createdContestAddress, true);
-  };
 
   return (
     <CWPageLayout>
@@ -30,16 +26,9 @@ const ContestLiveStep = () => {
             Your contest is now active at the address below. You can now add
             funds to the contest at any time.
           </CWText>
-          <div className="copy-address-input">
-            <CWText type="b2" className="address" fontWeight="medium">
-              {createdContestAddress}
-            </CWText>
-            <CWIconButton
-              iconName="copyNew"
-              buttonSize="med"
-              onClick={handleCopyLink}
-            />
-          </div>
+
+          <CopyAddressInput address={createdContestAddress} />
+
           <div className="buttons">
             <CWButton
               containerClassName="cta-btn"
